requirements.txt
git+https://github.com/espnet/espnet.git@3856d998ee0b2fa20f7b8fa48553754f33ed6e63
espnet_model_zoo
pyopenjtalk
datasets
torchaudio
typeguard==4.4.0
jiwer
fastapi
uvicorn
fugashi
pykakasi
basic-pitch[onnx]
audiobox_aesthetics
transformers
zhconv
git+https://github.com/sea-turt1e/kanjiconv
soundfile
PyYAML
gradio
google-generativeai
